resolved all pathing issues after file relocation
