+++ scribemaster/src/pages/CharacterInsertionPage.tsx
@@ -1,4 +1,3 @@
-import { CampaignExplorer2 } from "@/components/CampaignCreationComponents/CampaignExplorer2";
 import CharacterInsertion from "@/components/Character/AssetCreation/CharacterInsertion";
 import { SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar";
 import { AppSidebar } from "@/components/UtilityComponents/AppSidebar";
@@ -8,7 +7,7 @@ const CharacterInsertionPage = () => {
     <SidebarProvider>
       <AppSidebar />
       <SidebarTrigger />
-      <CampaignExplorer2 campaignId={1} />
+      <CharacterInsertion />
     </SidebarProvider>
   );
 };
